requirements.txt
absl-py==1.0.0
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.6.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
backoff==1.10.0
bcrypt==3.2.2
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect==1.1.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator==5.1.0
DLLogger @ git+https://github.com/NVIDIA/dllogger@0540a43971f4a8a16693a9de9de73c1072020769
einops==0.4.1
fake-useragent==0.1.11
fastapi==0.78.0
ffmpeg-normalize==1.23.0
ffmpeg-progress-yield==0.2.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.7.1
fonttools==4.25.0
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
gradio==3.0.23
grpcio==1.44.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.8.1
idna==3.3
imageio==2.13.5
importlib-metadata==4.11.2
inflect @ file:///C:/ci/inflect_1614797407862/work
Jinja2==3.1.2
joblib==1.1.0
kiwisolver @ file:///C:/ci/kiwisolver_1644962567532/work
librosa==0.8.1
linkify-it-py==1.0.3
llvmlite==0.37.0
lxml==4.8.0
Markdown==3.3.6
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib @ file:///C:/ci/matplotlib-suite_1647423638658/work
mdit-py-plugins==0.3.0
mdurl==0.1.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
monotonic==1.6
multidict==6.0.2
munkres==1.1.4
networkx==2.6.3
ninja==1.10.2.3
numba==0.54.1
numpy==1.20.3
oauthlib==3.2.0
opencv-python==4.5.4.60
orjson==3.7.6
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
paramiko==2.11.0
parse==1.19.0
Pillow==9.0.1
pooch==1.5.2
progressbar==2.5
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
pydub==0.25.1
pyee==8.2.2
PyNaCl==1.5.0
pynvml==11.4.1
pyparsing==3.0.6
pyppeteer==1.0.2
pyquery==1.4.3
pyTelegramBotAPI==4.7.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.2.0
PyYAML==6.0
regex==2022.6.2
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.1
resampy==0.2.2
rfc3986==1.5.0
rotary-embedding-torch==0.1.5
rsa==4.8
scenedetect==0.6.0.3
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
starlette==0.19.1
telebot==0.0.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.0.0
tifffile==2021.11.2
tokenizers==0.12.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.62.3
transformers==4.20.1
typing_extensions==4.0.1
uc-micro-py==1.0.1
Unidecode==1.3.4
urllib3==1.26.7
uvicorn==0.18.2
w3lib==1.22.0
websockets==10.2
Werkzeug==2.1.0
wincertstore==0.2
yarl==1.7.2
youtube-dl==2021.12.17
zipp==3.7.0
